+++ src/main/java/it/wirge/data/services/ISeedService.java
@@ -2,11 +2,9 @@ package it.wirge.data.services;
 
 import it.wirge.rest.models.Seed;
 
-/**
- * Created by enrico on 20/11/15.
- */
+
 public interface ISeedService {
-    Seed getSeed(int idSeed);
+    Seed getSeed(Long idSeed);
 
     Seed insertSeed(Seed seed) throws Exception;
